+++ lazybe/src/page.rs
@@ -21,7 +21,7 @@ impl PaginationInput {
 pub struct Page<T> {
     /// 0-index page number (0 is the first page)
     pub page: u32,
-    pub page_size: u64,
-    pub total_records: u64,
+    pub page_size: u32,
+    pub total_records: u32,
     pub data: Vec<T>,
 }
